Refs #2 - Rename getSprite to getNew and CarSprite to Car.
