+++ application/src/examples/car-driving-state.js
@@ -71,7 +71,7 @@ CarDrivingState.prototype.create = function()
 
     this.game.physics.p2.updateBoundsCollisionGroup();
 
-    this.car = this.carFactory.getSprite(this.game.world.centerX, this.game.world.centerY, 'car');
+    this.car = this.carFactory.getNew(this.game.world.centerX, this.game.world.centerY, 'car');
     this.game.world.addChild(this.car);
 
     this.car.body.setCollisionGroup(this.collisionGroup);
